requirements.txt
fastapi
uvicorn[standard]
torch
transformers
diffusers
Pillow
opencv-python
python-multipart
accelerate
